throttled call to processRequest during rate limiter setup rather than express middleware
